Create sub-reducers and course selector
Remove 'courses' from state tree.
Courses are calculated by the course selector.
Create tests for course selector.
Split 'reducer.js' into 'reducers/course.js' and 'reducers/filter.js'.
Modify reducer tests to fit the new structure.
