+++ package.json
@@ -45,7 +45,8 @@
     "react-dom": "15.1.0",
     "react-redux": "4.4.5",
     "react-router": "2.4.1",
-    "redux": "3.5.2"
+    "redux": "3.5.2",
+    "reselect": "2.5.1"
   },
   "devDependencies": {
     "autoprefixer": "6.3.6",
