+++ src/reducers/reducer.js
@@ -1,0 +1,8 @@
+import {combineReducers} from 'redux';
+import filter from './filter';
+import allCourses from './course';
+
+export default combineReducers({
+  allCourses,
+  filter
+});
